Add repository list view
