+++ app/assets/javascripts/sidebar_toggle.js.erb
@@ -36,7 +36,7 @@
       }
 
       $('.navbar-secondary').one("transitionend",function(event) {
-        $('.navbar-secondary').trigger(sideBarEvent);
+        $('#sidebar-wrapper').trigger(sideBarEvent);
       })
     }
