+++ app/assets/stylesheets/themes/scinote.scss
@@ -74,12 +74,6 @@ table {
   padding-top: 50px;
 }
 
-.content-pane {
-  background-color: $color-white;
-  margin: 20px 0;
-  padding: 25px 20px;
-}
-
 .spacer {
   margin-left: 0.5em;
   margin-right: 0.5em;
