+++ app/views/repositories/show.html.erb
@@ -128,7 +128,6 @@
 <%= javascript_include_tag 'repositories/edit' %>
 <%= javascript_include_tag 'repositories/repository_datatable' %>
 <%= javascript_include_tag "repositories/show" %>
-<%= javascript_include_tag "repositories/index" %>
 <%= javascript_pack_tag 'custom/inputmask' %>
 <%= javascript_pack_tag 'emoji_button' %>
